+++ src/main/java/net/silentchaos512/gear/setup/GearItemSets.java
@@ -24,7 +24,7 @@ public class GearItemSets {
     public static final GearItemSet<GearMacheteItem> MACHETE = set(GearTypes.MACHETE, "machete_blade", GearMacheteItem::new);
     public static final GearItemSet<GearDaggerItem> KNIFE = set(GearTypes.KNIFE, "knife_blade", GearDaggerItem::new);
     public static final GearItemSet<GearDaggerItem> DAGGER = set(GearTypes.DAGGER, "dagger_blade", GearDaggerItem::new);
-    public static final GearItemSet<GearSpearItem> SPEAR = set(GearTypes.SPEAR, "spear_tip", GearSpearItem::new);
+    public static final GearItemSet<GearSpearItem> SPEAR = set(GearTypes.SPEAR, "spearhead", GearSpearItem::new);
     public static final GearItemSet<GearTridentItem> TRIDENT = set(GearTypes.TRIDENT, "trident_prongs", GearTridentItem::new);
     public static final GearItemSet<GearMaceItem> MACE = set(GearTypes.MACE, "mace_core", GearMaceItem::new);