+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased 1.21.11]
+### Changed
+- Spears now work like vanilla spears
+- Spear Tip renamed to Spearhead
+
 ## [1.21.1-4.0.30] - 2025-11-23
 ### Added
 - The crude mixer, an early game alternative to the super mixer. It does that same thing (makes an alloy out of any two to four materials) but produces a weaker result.